fix syntax error, manually casting ints

+++ hydrapaper/wallpaper_merger.py
@@ -26,14 +26,15 @@ def multi_setup_pillow(monitors, save_path, wp_setter_func=None):
     offsets_y = []
     for m in monitors:
         if m.scaling == highest_scaling:
-            resolutions.append((m.width/highest_scaling, m.height/highest_scaling))
-            widths.append(m.width/highest_scaling)
-            heights.append(m.height/highest_scaling)
-            offsets.append((m.offset_x/highest_scaling, m.offset_y/highest_scaling))
-            offsets_x.append(m.offset_x/highest_scaling)
-            offsets_y.append(m.offset_y/highest_scaling)
+            resolutions.append((int(m.width/highest_scaling), int(m.height/highest_scaling)))
+            widths.append(int(m.width/highest_scaling))
+            heights.append(int(m.height/highest_scaling))
+            offsets.append((int(m.offset_x/highest_scaling), int(m.offset_y/highest_scaling)))
+            offsets_x.append(int(m.offset_x/highest_scaling))
+            offsets_y.append(int(m.offset_y/highest_scaling))
         else:
-            resolutions.append((m.width*highest_scaling, m.height*highest_scaling))            widths.append(m.width*highest_scaling)
+            resolutions.append((m.width*highest_scaling, m.height*highest_scaling))
+            widths.append(m.width*highest_scaling)
             heights.append(m.height*highest_scaling)
             offsets.append((m.offset_x*highest_scaling, m.offset_y*highest_scaling))
             offsets_x.append(m.offset_x*highest_scaling)
